bunch of cleanups and fixes to my fragments

+++ fragments/fastboot_image.sh
@@ -10,31 +10,23 @@
 export ROOT_FS_LABEL="Armbian_root"
 
 ## Hooks
-user_config__disable_all_image_fingerprints_and_compression() {
-	display_alert "Custom config stage" "user_config__disable_all_image_fingerprints_and_compression" "info"
+user_config__900_disable_all_image_fingerprints_and_compression() {
 	export COMPRESS_OUTPUTIMAGE=none_at_all_hopefully
 }
 
 # the real images have been produced elsewhere (.fastboot.boot and .fastboot.rootfs .imgs)
-post_build_image___discard_full_image() {
-	display_alert "Custom config stage" "post_build_image___discard_full_image" "info"
-	[[ -f "${FINALDEST}/${version}.img" ]] && {
-		rm -f "${FINALDEST}/${version}.img"
-		rm -f "${FINALDEST}/${version}.img.txt"
-	}
+post_build_image__discard_full_image() {
+	[[ -f "${FINALDEST}/${version}.img" ]] && rm -f "${FINALDEST}/${version}.img"
+	[[ -f "${FINALDEST}/${version}.img.txt" ]] && rm -f "${FINALDEST}/${version}.img.txt"
 }
 
 # last chance to modify mkopts and such, add labels to partitions
 prepare_partitions_custom__add_rootfs_label_to_mkfs() {
-	display_alert "Custom config stage" "prepare_partitions_custom" "info"
 	display_alert "prepare_partitions_custom oneplus adding to mfks ext4" "Label: ${ROOT_FS_LABEL}" "info"
 	mkopts[ext4]="-L ${ROOT_FS_LABEL} ${mkopts[ext4]}"
 }
 
-
 config_pre_umount_final_image__androidfastboot_extract_kernel() {
-	display_alert "Custom config stage" "config_pre_umount_final_image__androidfastboot_extract_kernel" "info"
-
 	# fastboot boot wants the gzipped kernel and concatenated dtb at the end. don't ask questions.
 	gzip -9 --keep --no-name "$MOUNT"/boot/vmlinuz-*
 	cat "$MOUNT"/boot/vmlinuz-*.gz "$MOUNT"/boot/dtb/"${BOOT_FDT_FILE}" >"$MOUNT"/boot/vmlinuz.gz.dtb
@@ -44,13 +36,11 @@ config_pre_umount_final_image__androidfastboot_extract_kernel() {
 	create_fastboot_boot_img "boot" "$MOUNT"/boot/initrd.img-* "root=LABEL=${ROOT_FS_LABEL} console=ttyGS0,115200 console=tty1 ${deviceinfo_kernel_cmdline}"
 }
 
-config_post_umount_final_image___extract_pure_ext4_image_from_partitioned_loop() {
-	display_alert "Custom config stage" "config_post_umount_final_image___extract_pure_ext4_image_from_partitioned_loop" "info"
+config_post_umount_final_image__extract_pure_ext4_image_from_partitioned_loop() {
 	local wanted_partition="${LOOP}p1"
 	local dest_img_file="${DEST}/images/${version}.fastboot.rootfs.img"
 	pv -N "[ .... ] dd" "${wanted_partition}" >"${dest_img_file}"
 }
-
 
 ## internal functions
 
@@ -78,5 +68,3 @@ create_fastboot_boot_img() {
 	display_alert "Android boot.img for fastboot" "${id}: .fastboot.${id}.img suffix" "info"
 
 }
-
-
